Make the server serve the client from http://localhost:9090/
This change will make the server serve the client code instead of manually having to start the client. It is done by adding our own request handler which determines if the request is for the piratesApi or not. If it is for the root we redirect to the piratebadge.html file. Otherwise we try to serve the request from the virtual diretory.

The steps for 2-6 are now.

1. run 'pub build' to generate the build/web directory with the client code
2. start the server
3. goto http://localhost:9090/ which should load the client

R=shaza@google.com
BUG=

Review URL: https://codereview.chromium.org/1068923003