Fix a few bugs in the codelab and change to port 8080 for the server.
I have fixed a bug where trying to kill a pirate that was not in the first part of the list would fail. I have also added an error if trying to add the same pirate twice (before it silently ignored it).
Also changed the truePirate method to check for null or empty name.
Finally changed the client code to display the errors it receives when calling the API (if any).

The reason for changing to port 8080 is that it is the port used on AppEngine plus it is a more common default than 9090.

Make the server serve the client from http://localhost:9090/

This change will make the server serve the client code instead of manually having to start the client. It is done by adding our own request handler which determines if the request is for the piratesApi or not. If it is for the root we redirect to the piratebadge.html file. Otherwise we try to serve the request from the virtual diretory.

The steps for 2-6 are now.

1. run 'pub build' to generate the build/web directory with the client code
2. start the server
3. goto http://localhost:9090/ which should load the client

R=shaza@google.com
BUG=

Review URL: https://codereview.chromium.org/1080313002